PFP-9075 fikset review kommentar og implementerte funkjonalitet å lukke oppgaver når behandling fristen endres
